+++ HLS/baseline_implementation/max_stall_logic_throughput_tests.py
@@ -2,7 +2,7 @@
 import subprocess
 
 OPERATIONS = [(0,"add"),(1,"mul"),(2,"div")]
-TYPES = [(0,"double"),(1,"float"),(2,"int32")]
+TYPES = [(0,"double"),(1,"float"),(2,"int64")]
 
 DD_configs = [(1.5,8),(1.25,12),(1.25,16),(1.25,32)]
 OP_CODE = 0
